+++ SpatialGDK/Source/SpatialGDK/Public/SpatialConstants.h
@@ -203,6 +203,10 @@ namespace SpatialConstants
 	const FString SPATIALOS_METRICS_DYNAMIC_FPS = TEXT("Dynamic.FPS");
 
 	const FString LOCATOR_HOST = TEXT("locator.improbable.io");
+	// URL that can be used to reconnect using the command line arguments.
+	const FString RECONNECT_USING_COMMANDLINE_ARGUMENTS = TEXT("0.0.0.0");
+	const FString URL_LOGIN_OPTION = TEXT("login=");
+	const FString URL_PLAYER_IDENTITY_OPTION = TEXT("playeridentity=");
 	const uint16 LOCATOR_PORT = 443;
 
 	const FString DEVELOPMENT_AUTH_PLAYER_ID = TEXT("Player Id");